Security fixes and features
Context separation is a feature of the managed permission map that essentially passes copies of the actual permission map to the resolvers.
This feature adds a performance overhead as the values have to be copied for each resolution.

Resolver trust is a similar feature that makes sure that the permissions requested by resolvers are proper permissions.
This feature is not particularly effective without context separation.

Also, the ESM tsconfig has to use Node module resolution as otherwise it does not resolve packages properly.